% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.names}
\alias{filter.names}
\title{Filter a subset from the names based on percentiles}
\usage{
filter.names(...)
}
\arguments{
\item{...}{Predicates on the dataset using non-standard evaluation}
}
\description{
This method filters a set of names from the complete
set. The set is filtered based on predicates, which
are evaluated in relation to the percentiles of the
given variable (see examples). Multiple predicates
are first all evaluated on the complete dataset and
then combined (i.e. the percentiles are not
re-evaluated in the filtered dataset).
}
\examples{

# Retrieve all names without any further reduction
filter.names()

# Retrieve all names with Intelligence ratings above the median
filter.names(Intelligence >= 0.5)

# Retrieve all names, which are not too uncommon (Familiarity at least
# in the 20th percentile) and for which Intelligence is above the
# median
filter.names(Familiarity >= 0.2, Intelligence >= 0.5)

}
