% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match.groups}
\alias{match.groups}
\title{Create a group of names all similar to each other}
\usage{
match.groups(split, n, discard = 0, subset = filter.names(),
  ga.params = list(), ...)
}
\arguments{
\item{split}{The name of the rating on which the split
should be performed}

\item{n}{Number of names from each group to select}

\item{discard}{The percentage of names too close to the median
which should be discarded (default: 0, i.e. keep
all names)}

\item{subset}{An optional subset on which the split should be
done. If this is left out, the split will
be created on all names.}

\item{ga.params}{Parameters to genetic algorithm
\describe{
   \item{penalty.factor}{How strongly to penalize candidate solution,
    which do not have the right number of names. Increase, if the wrong
    number of names is returned repeatedly}
   \item{init.factor}{How many more names to include initialy to increase
    the search space.}
   \item{maxiter}{Maximal number of iterations. See \code{\link[GA]{ga}} for details.}
   \item{run}{Number of consecutive generations without improvement. See \code{\link[GA]{ga}}
    for details.}
   \item{popSize}{Initial population size. See \code{\link[GA]{ga}} for details.}
   \item{pmutation}{Mutation probability. See \code{\link[GA]{ga}} for details.}
   \item{pcrossover}{Crossover probability. See \code{\link[GA]{ga}} for details.}
   \item{elitism}{Number of best candidates to keep each iteration.
    See \code{\link[GA]{ga}} for details.}
   \item{parallel}{Should the GA run in parallel? See \code{\link[GA]{ga}} for details.}
   \item{monitor}{Monitor function. See \code{\link[GA]{ga}} for details.}
}}

\item{...}{Arguments passed on to \code{names.dist}
\describe{
  \item{x}{The object containing the names}
}}
}
\description{
Creates a group of names with n members from two groups, such that all names
are as similar to each other as possible. Names are not only similar across groups
but also to each other in the group.
}
\details{
Uses a genetic algorithm to find the best names, so different results may be
found each run.

To seed the genetic algorithm with well suited sets, first random sets of names
are created. To increase the search space, more names are selected than will later
be retrieved for each of the partitions. The number of additional names, which
are included is controlled by a ga.param value.
}
