% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{names.dist.names.selection}
\alias{names.dist.names.selection}
\title{Calculate the distance between all pairs of names in a set}
\usage{
\method{names}{dist.names.selection}(x, ...)
}
\arguments{
\item{x}{The object containing the names}

\item{...}{Further weighting of attributes (see examples)}
}
\description{
Calculate the distance between all pairs of names in a set
}
