% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{names.dist.names.split}
\alias{names.dist.names.split}
\title{Calculate the distance between all names from two groups of names}
\usage{
\method{names}{dist.names.split}(x, ...)
}
\arguments{
\item{x}{The object containing the names}

\item{...}{Further weighting of attributes (see examples)}
}
\description{
Calculate the distance between all names from two groups of names
}
