% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{names.dist}
\alias{names.dist}
\title{Calculate distance between sets of names}
\usage{
names.dist(x, ...)
}
\arguments{
\item{x}{The object containing the names}

\item{...}{Further weighting of attributes (see examples)}
}
\description{
Calculate distance between sets of names
}
