% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{names.long}
\alias{names.long}
\title{Validated ratings for German names (long version)}
\format{A data frame with 39017 rows and 49 variables:}
\source{
\url{https://osf.io/jepzp/}
}
\usage{
names.long
}
\description{
This dataset contains the individual ratings of all participants
for all of the rated names. See \url{https://osf.io/jepzp/} for
a complete description.
}
\keyword{datasets}
