% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{names.mean}
\alias{names.mean}
\title{Validated ratings for German names (averaged version)}
\format{A data frame with 2000 rows and 50 variables:}
\source{
\url{https://osf.io/jepzp/}
}
\usage{
names.mean
}
\description{
This dataset contains the ratings for all 2000 names averaged
across all participants who rated the same name.
See \url{https://osf.io/jepzp/} for a complete description.
}
\keyword{datasets}
