% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings.R
\name{names.ratings}
\alias{names.ratings}
\title{Show the ratings for a set of names}
\usage{
\method{names}{ratings}(..., subset = filter.names())
}
\arguments{
\item{...}{The ratings that should be retrieved}

\item{subset}{The subset for which the ratings should be retrieved}
}
\description{
Show the ratings for a set of names
}
\examples{

s <- filter.names(Intelligence > 0.75)

names.ratings(subset=s, Intelligence, Competence)

}
