% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match.pairs}
\alias{match.pairs}
\title{Match most similar pairs for trial generation in a subset of names}
\usage{
match.pairs(n1, n2, ...)
}
\arguments{
\item{n1}{First subset of the names used in pairing}

\item{n2}{Second subset of the names used in pairing}

\item{...}{Arguments passed on to \code{match.split}
\describe{
  \item{s}{The split of the names. Only the first two groups will be used,
if the split contains more than two groups.}
}}
}
\description{
This function matches pairs of names based on similarity to create
a set of trials. Names are matched, such that each name is only used
once and the sum of the distances over all names in total is minimized.
}
\details{
Normally the principal components are used for matching names, but
other ratings can also be added and emphasized to create sets
of names which are more similar along a special set of ratings.

Note: This function takes the two groups of names as separate arguments,
}
\examples{

# Split the names according to intelligence
s <- partition.names(Intelligence)

# Further split the names randomly
low <- partition.names.random(subset=s[1], prop=c(1,1,1,1))
high <- partition.names.random(subset=s[1], prop=c(1,1,1,1))

# Create sets with matching high/low intelligence names
low_low <- match.pairs(low[1], low[2])
low_high <- match.pairs(low[3], high[1])
high_low <- match.pairs(high[2], low[4])
high_high <- match.pairs(high[3], high[4])

}
