% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match.pairs}
\alias{match.pairs}
\title{Match most similar pairs for trial generation in a subset of names}
\usage{
match.pairs(split, discard = 0, subset = filter.names(), ...)
}
\arguments{
\item{split}{The name of the rating on which the split
should be performed}

\item{discard}{The percentage of names too close to the median
which should be discarded (default: 0, i.e. keep
all names)}

\item{subset}{An optional subset on which the split should be
done. If this is left out, the split will
be created on all names.}

\item{...}{Arguments passed on to \code{names.dist}
\describe{
  \item{x}{The object containing the names}
}}
}
\description{
Match most similar pairs for trial generation in a subset of names
}
